requirements.txt
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.2.1
cmake==3.25.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.4
Jinja2==3.1.4
joblib==1.5.1
lit==15.0.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pillow==11.0.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.28.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.54.1
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==1.26.13
uvicorn==0.35.0
